+++ tests/proud-core-stubs.php
@@ -7,3 +7,11 @@ function proud_html_preview_queue_cleanup( $provider, $artifact_key, $artifact_u
 
 	return true;
 }
+
+function proud_html_preview_url( $post_id, $source_url = '' ) {
+	$GLOBALS['filetoweb_test_preview_url_calls'][] = array( $post_id, $source_url );
+
+	return isset( $GLOBALS['filetoweb_test_preview_urls'][ $post_id ] )
+		? $GLOBALS['filetoweb_test_preview_urls'][ $post_id ]
+		: '';
+}